Function aggregate_sig::verify_aggregate_sig

pub fn verify_aggregate_sig<P: Pairing>(
    m_pk_pairs: &[(&[u8], PublicKey<P>)], 
    signature: Signature<P>
) -> bool

Verifies an aggregate signature on pairs (messages m_i, PK_i) for i=1..n by checking pairing(sig, g_2) == product_{i=0}^n ( pairing(g1_hash(m_i), PK_i) ) where g_2 is the generator of G2. Verification returns false if any two messages are not distinct

For now, the generator used is the default generator of the underlying library however, this should be parametrized in the future